backend/requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==21.4.0
autopep8==1.6.0
Babel==2.10.1
backports.zoneinfo==0.2.1
bcrypt==3.2.2
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
Django==4.0.4
django-cors-headers==3.12.0
django-dotenv==1.4.2
djangorestframework==3.13.1
h11==0.13.0
idna==3.3
outcome==1.1.0
Pillow==8.0.1
praw==7.6.0
prawcore==2.3.0
PyAutoIt==0.6.0
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
selenium==3.141.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
toml==0.10.2
trio==0.20.0
trio-websocket==0.9.2
update-checker==0.18.0
urllib3==1.26.9
webdriver-manager==3.7.0
websocket-client==1.3.2
wget==3.2
wsproto==1.1.0
